Issue #3208252: Inserting projects can fail on release data: simplytest_project_versions.date

+++ modules/simplytest_projects/src/ProjectVersionManager.php
@@ -62,7 +62,7 @@ final class ProjectVersionManager {
             'version' => $release->version,
             'tag' => $release->tag,
             'date' => $release->date,
-            'status' => $release->status === 'published',
+            'status' => (int) ($release->status === 'published'),
             'core_compatibility' => $release->core_compatibility,
           ])
           ->execute();
